statistics: refactor common functions into subscriber
